Refactor basic FHIR utility methods to appropriate scope for reuse.
